[VOGRE-9] fixed group_id in repository_views, added class in managers, removed extra file
